server/requirements.txt
Pyrebase==3.0.27
python_twitter==3.4.1
